backend/requirements.txt
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
joblib==1.5.2
numpy==2.3.3
pandas==2.3.2
scipy==1.15.3
lightgbm
